move python files to micro benchmark
